feat: ReportRepositoryに切り出した
