+++ modules/model/src/main/java/net/pantasystem/milktea/model/user/report/SendReportUseCase.kt
@@ -5,13 +5,12 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.withContext
 import net.pantasystem.milktea.common.runCancellableCatching
-import net.pantasystem.milktea.model.user.UserRepository
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
 class SendReportUseCase @Inject constructor(
-    val userRepository: UserRepository
+    private val reportRepository: ReportRepository,
 ){
     operator fun invoke(state: ReportState): Flow<ReportState> {
         return flow {
@@ -32,7 +31,7 @@ class SendReportUseCase @Inject constructor(
                         state.comment
                     )
                     runCancellableCatching {
-                        userRepository.report(
+                        reportRepository.create(
                             r
                         )
                     }.onSuccess {
